paginator

+++ src/main/java/com/akira/RO/Logs.java
@@ -3,11 +3,11 @@ package com.akira.RO;
 import java.util.List;
 
 public class Logs {
-private int pages;
-public int getPages() {
+private long pages;
+public long getPages() {
 	return pages;
 }
-public void setPages(int pages) {
+public void setPages(long pages) {
 	this.pages = pages;
 }
 public List getList() {
@@ -17,7 +17,7 @@ public void setList(List list) {
 	this.list = list;
 }
 private List list;
-public Logs(int pages,List list){
+public Logs(long pages,List list){
 	this.list=list;
 	this.pages=pages;
 }
